+++ lib/commands/run-test-scenario.ts
@@ -11,6 +11,7 @@ import {
   PostmanCollectionGeneratorOption,
 } from "../testScenario/postmanCollectionGenerator";
 import { inversifyGetInstance } from "../inversifyUtils";
+import { getApiVersionFromSwaggerFile, getProviderFromFilePath } from "../util/utils";
 import { getSwaggerFilePathsFromTestScenarioFilePath } from "./../testScenario/testResourceLoader";
 
 export const command = "run-test-scenario <test-scenario>";
@@ -27,7 +28,12 @@ export const builder: yargs.CommandBuilder = {
     alias: "outputDir",
     describe: "the output folder.",
     string: true,
-    default: "generated_collections",
+    default: "generated",
+  },
+  uploadBlob: {
+    describe: "upload generated collection to blob.",
+    boolean: true,
+    default: false,
   },
 };
 
@@ -46,19 +52,25 @@ export async function handler(argv: yargs.Arguments): Promise<void> {
     }
     // fileRoot is the nearest common root of all swagger file paths
     const fileRoot = path.dirname(swaggerFilePaths[0]);
+    const resourceProvider = getProviderFromFilePath(testScenarioFilePath);
+    const apiVersion = getApiVersionFromSwaggerFile(swaggerFilePaths[0]);
     const opt: PostmanCollectionGeneratorOption = {
-      name: testScenarioFilePath.replace(/^.*[\\\/]/, "").replace(".yaml", ""),
+      name: `${resourceProvider}/${apiVersion}/${testScenarioFilePath
+        .replace(/^.*[\\\/]/, "")
+        .replace(".yaml", "")}`,
       testDef: testScenarioFilePath,
       swaggerFilePaths: swaggerFilePaths,
       fileRoot: fileRoot,
       checkUnderFileRoot: false,
-      generateCollection: false,
+      generateCollection: true,
       useJsonParser: false,
       runCollection: true,
       env: env,
       outputFolder: argv.output,
       eraseXmsExamples: false,
       eraseDescription: false,
+      enableBlobUploader: argv.uploadBlob,
+      blobConnectionString: process.env.blobConnectionString || "",
     };
     const generator = inversifyGetInstance(PostmanCollectionGenerator, opt);
     await generator.GenerateCollection();