+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Change Log - oav
 
+## 04/14/2021 2.3.3
+
+- Support auto upload generated file to azure blob.
+- Support mask bearer token
+- generate-static-test-scenario support generate readme.test.md
+
 ## 04/06/2021 2.3.2
 
 - Support auto generate test scenario based on resource type.